Fix crosshair watch (#5)
There were a couple of problems with crosshair watch command.

First, the screen did not refresh properly, which was fixed in the
version 0.0.13 of crosshair.

Second, invoking crosshair as a module using `sys.executable` caused
crosshair not to verify the file. Once we execute the command just using
`!` in thonny terminal, the problem was resolved.